requirements.txt
langchain
langchain_community
langchain_core
langchain_groq
langchain_google_genai
langchain-text-splitters
langchainhub
python-docx
langchain-openai
pinecone-client
langchain-huggingface
rich
google-generativeai
python-dotenv
docx2txt
streamlit
pypdf
langchain-pinecone
edge_tts
streamlit_mic_recorder
langchain_cohere